idf_monitor: Fix console performance
https://github.com/espressif/esp-idf/commit/5b64c9cfbb40ede2f62774e3d1c783ffcbf28f5c
introduced filtering options which handles the last (unterminated) line
with a delay. This introduced poor performance for console applications
when the user interacts with the ESP32 device in the same line because
there is an artificial delay for each key-press.

https://github.com/espressif/esp-idf/commit/a7b52aa9574133650d8d7016d160c3e01c021422
decreased the delay but there are still delays for each key-presses.

The current fix makes only one delay per line instead of each
key-presses. When an unterminated line is detected then no more
"delayed" print is used for the given line.

Fixes https://github.com/espressif/esp-idf/issues/2054